+++ BrownieHound/BrownieHound/capture.xaml.cs
@@ -36,6 +36,7 @@ using Reactive.Bindings.Extensions;
 using Org.BouncyCastle.Asn1.X509;
 using Org.BouncyCastle.Bcpg;
 using static BrownieHound.RuleData;
+using static BrownieHound.ReadPacketData;
 
 namespace BrownieHound
 {
@@ -44,81 +45,6 @@ namespace BrownieHound
     /// </summary>
     public partial class capture : Page
     {
-        public class packetData
-        {
-            public int Number { get; set; }
-            public DateTime Time { get; set; }
-            public string Source { get; set; }
-            public string Destination { get; set; }
-            public string sourcePort { get; set; }
-            public string destinationPort { get; set; }
-            public string Protocol { get; set; }
-            public int frameLength { get; set; }
-            public string Info { get; set; }
-            public string Data { get; set; }
-            List<string> protocols = new List<string>();
-
-            public packetData(String err)
-            {
-                Info = err;
-                
-            }
-            public packetData(JObject layersObject)
-            {
-                Data = JsonConvert.SerializeObject(layersObject, Newtonsoft.Json.Formatting.None);
-                foreach (var layer in layersObject)
-                {
-                    protocols.Add(layer.Key.ToString());
-                    //Debug.WriteLine(layer.Key);
-                }
-                Number = Int32.Parse((string)layersObject[protocols[0]][$"{protocols[0]}_{protocols[0]}_number"]);
-                //frame_frame_number
-
-                string caputureTime = (string)layersObject[protocols[0]][$"{protocols[0]}_{protocols[0]}_time"];
-                caputureTime = caputureTime.Substring(0, 27);
-                //精度が高すぎるので落とす
-
-                Time = DateTime.ParseExact(caputureTime, "yyyy-MM-dd'T'HH:mm:ss.FFFFFFF", CultureInfo.InvariantCulture, DateTimeStyles.RoundtripKind);
-                Time = Time.AddHours(9);
-
-
-                string eSource = (string)layersObject[protocols[1]][$"{protocols[1]}_{protocols[1]}_src"];
-                string eDestination = (string)layersObject[protocols[1]][$"{protocols[1]}_{protocols[1]}_dst"];
-                //ethレベルのアドレス（MACアドレス）
-
-                Source = (string)layersObject[protocols[2]][$"{protocols[2]}_{protocols[2]}_src"];
-                Destination = (string)layersObject[protocols[2]][$"{protocols[2]}_{protocols[2]}_dst"];
-
-                if (Source == null)
-                {
-                    Source = eSource;
-                    Destination = eDestination;
-                }
-                if (protocols.Contains("tcp") || protocols.Contains("udp"))
-                {
-                    sourcePort = (string)layersObject[protocols[3]][$"{protocols[3]}_{protocols[3]}_srcport"];
-                    destinationPort = (string)layersObject[protocols[3]][$"{protocols[3]}_{protocols[3]}_dstport"];
-                    Info += $"{sourcePort} → {destinationPort}";
-                }
-
-                if (protocols.Last().Equals("data"))
-                {
-                    Protocol = protocols[protocols.Count - 2];
-                }
-                else
-                {
-                    Protocol = protocols.Last();
-                }
-                Protocol = Protocol.ToUpper();
-
-                frameLength = Int32.Parse((string)layersObject[protocols[0]][$"{protocols[0]}_{protocols[0]}_len"]);
-                Info += $" {protocols.Last()}";
-                //Debug.WriteLine($"{Number} : {time.TimeOfDay} : {Source} : {Destination} : {Protocol} : {Length} :: {Info}");
-
-            }
-
-        }
-
         Process processTscap = null;
         string tsInterfaceNumber = "";
         List<DispatcherTimer> detectTimerList = new List<DispatcherTimer>();
@@ -413,23 +339,6 @@ namespace BrownieHound
 
             clock++;
             GC.Collect();
-        }
-        private packetData transfar(string msg)
-        {
-            packetData pd = null;
-            try
-            {
-                JObject packetObject = JObject.Parse(msg);
-                if (packetObject["layers"] != null)
-                {
-                    pd = new packetData((JObject)packetObject["layers"]);
-                }
-                }
-            catch
-            {
-                pd = (new packetData(msg));
-            }
-            return pd;
         }
 
         private void detectLogic(int detectionNumber)
@@ -799,7 +708,7 @@ namespace BrownieHound
             }
             for (int i = 0; i < 500; i++)
             {
-                CaptureData.Items.Insert(i, transfar(viewPacketStrings[i]));
+                CaptureData.Items.Insert(i, transfer(viewPacketStrings[i]));
             }
             if (CaptureData.Items.Count > 1000)
             {
@@ -845,7 +754,7 @@ namespace BrownieHound
             }
             for (int i = 0; viewPacketStrings[i] != null; i++)
             {
-                CaptureData.Items.Add(transfar(viewPacketStrings[i]));
+                CaptureData.Items.Add(transfer(viewPacketStrings[i]));
             }
             if (CaptureData.Items.Count > 1000)
             {
@@ -944,7 +853,7 @@ namespace BrownieHound
 
             for (int i = 0;viewPacketStrings[i] != null;i++)
             {
-                CaptureData.Items.Add(transfar(viewPacketStrings[i]));
+                CaptureData.Items.Add(transfer(viewPacketStrings[i]));
             }
             viewPacketStrings = null;
             GC.Collect();
@@ -990,7 +899,7 @@ namespace BrownieHound
                     }
                     for (int i = 0; viewPacketStrings[i] != null; i++)
                     {
-                        CaptureData.Items.Add(transfar(viewPacketStrings[i]));
+                        CaptureData.Items.Add(transfer(viewPacketStrings[i]));
                     }
                     viewPacketStrings = new string[501];
                     GC.Collect();
@@ -1017,7 +926,7 @@ namespace BrownieHound
                 }
                 for (int i = 0; viewPacketStrings[i] != null; i++)
                 {
-                    CaptureData.Items.Add(transfar(viewPacketStrings[i]));
+                    CaptureData.Items.Add(transfer(viewPacketStrings[i]));
                 }
                 viewPacketStrings = null;
                 GC.Collect();
